move operations after confirmation query, fix, tested

+++ backup-Root.sh
@@ -15,7 +15,7 @@ fi
 BackupPath="/Data/Backup/Linux/NVMeRoot"	#defines default BackupPath
 source=/									#defines backup source
 
-while getopts ":p:e:f:l:s:" arg; do
+while getopts ":p:e:f:ls:" arg; do
   case $arg in
     p)
       BackupPath=${OPTARG}
@@ -30,12 +30,11 @@ while getopts ":p:e:f:l:s:" arg; do
       source=${OPTARG}
       ;;
 	l)
-	  lvmSource="$(df --output=source / | tail -1)"
-	  lvm="$(basename "$lvm")-$(date +%F-%0H-%0M)"
-	  source="$(mktemp -d /mntXXX)"
-	  lvcreate -L 5G --snapshot -n "$lvm" "$lvmSource" \
-	  && mount /dev/mapper/"$lvm" "$source" \
-	  || exit 1
+	  lvm="$(df --output=source "$source" | tail -1)"	#e.g. /dev/mapper/ssd-root
+	  lvm_snap="$(basename "$lvm")"						#-> ssd-root
+	  lvm_VG="${lvm%%-*}"								#-> /dev/mapper/ssd
+	  lvm_snap="${lvm#*-}-snap-$(date +%F-%0H-%0M)"		#-> root-snap-d2018-03-16-13-31
+	  source="$(mktemp -d /media/temp.XXX)" || exit 1
 	  ;;
     \?)
       Print-Usage
@@ -47,22 +46,31 @@ shift $((OPTIND-1))
 if [ "$(df --output=source / | tail -1)" == "$(df --output=source "$BackupPath" | tail -1)" ]; then
 	printf "Warning: BackupPath (\"%s\") is on the same Filesystem as the Root Filesystem (\"/\")\\n" "$BackupPath"
 fi
-printf "BackupPath = \"%s\"\\n" "$BackupPath" >&2
-printf "RemoteShell = \"%s\"\\n" "$RemoteShell" >&2
-printf "ExcludeFiles = \"%s\"\\n" "$ExcludeFiles" >&2
+printf	"${lvm:+lvm mapping: \"$lvm\"\\n}\
+source = \"%s\"
+BackupPath = \"%s\"
+RemoteShell = \"%s\"
+ExcludeFiles = \"%s\"\\n" \
+"$source" "$BackupPath" "$RemoteShell" "$ExcludeFiles" >&2
+
 read -r -n 1 -p "Do you want to continue? [Y/n] " ans || { echo; exit; } && echo
 case "$ans" in
     [yY][eE][sS]|[yY]|"")
-        ;;
+      ;;
     *)
-
-	exit
-        ;;
+      if test "$lvm"; then
+		  rm -d "$source"
+	  fi
+      ;;
 esac
 
+if test "$lvm"; then
+	lvcreate -L 5G --snapshot -n "$lvm_snap" "$lvm" \
+	&& mount "$lvm_VG-${lvm_snap//-/--}" "$source" || exit 1
+fi
 rsync ${RemoteShell:+-e "$RemoteShell"} -i -ahxHAX --delete ${ExcludeFiles:+"--exclude-from=$ExcludeFiles"} / "$BackupPath"/
 if test "$lvm"; then
-	umount "$lvm"
-	lvremove -f "$source"
-	rm -d source
+	umount "$source"
+	lvremove -f "$(basename "$lvm_VG")/$lvm_snap"
+	rm -d "$source"
 fi
